Move run_spdk logic from main.rs of rust_wrapper example to spdk-rs; delete code from starfish; and remove all the dependencies on the future-preview
