+++ examples/hello_nvme_bdev_rust_wrapper/src/main.rs
@@ -39,27 +39,6 @@ use std::env;
 //use futures::FutureExt;
 //use tokio_core::reactor::Core;
 
-use std::future::{Future as StdFuture};
-use tokio::runtime::Runtime;
-use tokio::prelude::*;
-
-async fn map_ok<T: StdFuture>(future: T) -> Result<(),()> {
-    let _ = await!(future);
-    Ok(())
-}
-
-pub fn run_spdk<F>(future: F)
-where F: StdFuture<Output = ()> + Send + 'static,
-{
-    use tokio_async_await::compat::backward;
-    let future = backward::Compat::new(map_ok(future));
-
-    let mut rt = Runtime::new().unwrap();
-    rt.block_on(future);
-    rt.shutdown_now().wait().unwrap();
-    spdk_rs::event::app_stop(true);
-}
-
 async fn run() {
     match await!(run_inner()) {
             Ok(_) => println!("Successful"),
@@ -120,7 +99,7 @@ fn main()
     let ret = opts.start(|| {
         //NOTE: Alternatively, we can use `tokio::run_async(run())` but doing so requires us to use
         // C-c to terminate program
-        run_spdk(run());
+        spdk_rs::run::run_spdk(run());
     });
     println!("Successfully shutdown SPDK framework");
 }
